src/backend_test_deploy/requirements.txt
fastapi
uvicorn
python-dotenv
mysql-connector-python
google-cloud-storage
python-multipart
baml-py==0.218.1